+++ packages/core/index.ts
@@ -6,7 +6,6 @@
 
 export * from "./constants.js";
 export * from "./events.js";
-export * from "./sandbox.js";
 export * from "./utils.js";
 export * from "./model-cache.js";
 export * from "./tui-width.js";